+++ lib/service/Groups.dart
@@ -147,7 +147,7 @@ class Groups /* with Service */ {
     return null;
   }
 
-  Future<String> createGroup(Group group) async {
+  Future<GroupError> createGroup(Group group) async {
     if(group != null) {
       String url = '${Config().groupsUrl}/groups';
       try {
@@ -158,36 +158,52 @@ class Groups /* with Service */ {
         String body = AppJson.encode(json);
         Response response = await Network().post(url, auth: NetworkAuth.User, body: body);
         int responseCode = response?.statusCode ?? -1;
-        String responseBody = response?.body;
-        Map<String, dynamic> jsonData = ((responseBody != null) && (responseCode == 200)) ? AppJson.decodeMap(responseBody) : null;
-        if(jsonData != null){
-          String groupId = jsonData['inserted_id'];
-          NotificationService().notify(notifyGroupCreated, group.id);
-          return groupId;
+        if (responseCode == 200) {
+          Map<String, dynamic> jsonData = AppJson.decodeMap(response?.body);
+          String groupId = (jsonData != null) ? AppJson.stringValue(jsonData['inserted_id']) : null;
+          if (AppString.isStringNotEmpty(groupId)) {
+            NotificationService().notify(notifyGroupCreated, group.id);
+            return null; // succeeded
+          }
+        }
+        else if (responseCode == 400) {
+          Map<String, dynamic> jsonData = AppJson.decodeMap(response?.body);
+          Map<String, dynamic> jsonError = (jsonData != null) ? AppJson.mapValue(jsonData['error']) : null;
+          if (jsonError != null) {
+            return GroupError.fromJson(jsonError); // error descrition
+          }
         }
       } catch (e) {
         print(e);
       }
     }
-    return null;
+    return GroupError(); // generic error
   }
 
-  Future<bool> updateGroup(Group group) async {
+  Future<GroupError> updateGroup(Group group) async {
     if(group != null) {
       String url = '${Config().groupsUrl}/groups/${group.id}';
       try {
         Map<String, dynamic> json = group.toJson();
         String body = AppJson.encode(json);
         Response response = await Network().put(url, auth: NetworkAuth.User, body: body);
-        if((response?.statusCode ?? -1) == 200){
+        int responseCode = response?.statusCode ?? -1;
+        if(responseCode == 200){
           NotificationService().notify(notifyGroupUpdated, group.id);
-          return true;
+          return null;
+        }
+        else if (responseCode == 400) {
+          Map<String, dynamic> jsonData = AppJson.decodeMap(response?.body);
+          Map<String, dynamic> jsonError = (jsonData != null) ? AppJson.mapValue(jsonData['error']) : null;
+          if (jsonError != null) {
+            return GroupError.fromJson(jsonError); // error descrition
+          }
         }
       } catch (e) {
         print(e);
       }
     }
-    return false;
+    return GroupError(); // generic error
   }
 
   Future<bool> deleteGroup(String groupId) async {